general fixes from sparcur testing
Doi.ttl better handling of not actually ttl

cache.COOLDOWN to prevent retrying trying known bad ids

Pio get down the the string if a stream or identifier is provided,
might be worth implementing asStr or asBaseId or something like that.

Pio and Orcid fixes for cases where data/name could be None.