specify __hash__, __eq__, and id behavior for streams
Streams are objects that can be partially dereferenced and can have
multiple referents, they are essentially an augmented name. We do not
want the hash value of a stream to be dependent on its state since
that would mean that the same stream could be added to a dictionary
multiple times (among other things). Therefore we make hash and
equality a comparison between the class of the stream and the
identifier of the stream (or its resolved asStr identifier, a decision
that may be revisited since it seems unlikely to have externally
visible impacts unless someone fiddles with the asStr method
independent of __eq__ or __hash__).